Add files via upload
House Price Prediction Dataset is one of the most worked dataset to predict the price. This dataset gives very exposure to lot of aspects for exploratory data analysis. I have made my efforts to analyse as much as possible. Still there is scope for improvement in this file which will be addressed subsequently.